Rewrite admin command
